templates/rl/{{cookiecutter.project_name}}/requirements.txt
torch
hydra-core
numpy<2
gymnasium
pygame
tqdm